requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
psycopg[binary]>=3.1.18,<3.3
pyyaml>=6.0.1
python-dotenv>=1.0.1
tavily-python>=0.3.0
